+++ projectq/ops/__init__.py
@@ -23,18 +23,10 @@ from ._basics import (
     ClassicalInstructionGate,
     FastForwardingGate,
     BasicMathGate,
-    BasicPhaseGate
+    BasicPhaseGate,
 )
 from ._command import apply_command, Command, CtrlAll, IncompatibleControlState
-from ._metagates import (
-    DaggeredGate,
-    get_inverse,
-    is_identity,
-    ControlledGate,
-    C,
-    Tensor,
-    All
-)
+from ._metagates import DaggeredGate, get_inverse, is_identity, ControlledGate, C, Tensor, All
 from ._gates import *
 from ._qftgate import QFT, QFTGate
 from ._qubit_operator import QubitOperator